medakai/requirements.txt
streamlit==1.38.0
requests==2.32.3
python-dotenv==1.0.1
googletrans==4.0.0-rc1
twilio==9.3.0